+++ python/cudf/io/__init__.py
@@ -1,2 +1,4 @@
 # Copyright (c) 2018, NVIDIA CORPORATION.
-from .csv import read_csv
+
+from cudf.io.csv import read_csv
+from cudf.io.parquet import read_parquet